+++ app/store/slice/modalStore.ts
@@ -1,0 +1,33 @@
+import { create } from "zustand";
+import { persist } from "zustand/middleware";
+import { immer } from "zustand/middleware/immer";
+
+export type ModalState = {
+  isOpen: boolean;
+};
+
+export type ModalAction = {
+  actions: {
+    handleClose: () => void;
+    handleOpen: () => void;
+  };
+};
+
+export const useModalStore = create<ModalState & ModalAction>()(
+  persist(
+    immer((set) => ({
+      isOpen: false,
+      actions: {
+        handleClose: () =>
+          set((state) => {
+            state.isOpen = false;
+          }),
+        handleOpen: () =>
+          set((state) => {
+            state.isOpen = true;
+          }),
+      },
+    })),
+    { name: "todos-storage" },
+  ),
+);
